more MVStore -> FileStore moves: caches, chunks, executors/locks
